src/requirements.txt
fastapi==0.50.0
psycopg2-binary==2.8.4
PyGithub==1.46
python-twitter==3.5
requests==2.23.0
SQLAlchemy==1.3.13
uvicorn==0.11.3


# dev
pytest==5.3.5
pytest-mock==2.0.0
pytest-pycodestyle==2.0.0
sqlalchemy-utils==0.36.1
